Move final test to poll_api
